// src/shared/Brackets/types.tsx
export type Participant = {
    id: string | number;

    isWinner?: boolean;

    name?: string;

    status?: 'PLAYED' | 'NO_SHOW' | 'WALK_OVER' | 'NO_PARTY' | string | null;

    resultText?: string | null;

    [key: string]: any;
};

export type Match = {
    id: number | string;

    /** Link to this match. While onClick() can be used, providing an href
        better supports opening a new tab, or copying a link. * */
    href?: string;

    name?: string;

    nextMatchId: number | string | null;

    nextLooserMatchId?: number | string;

    tournamentRoundText?: string;

    startTime: string;

    state: 'PLAYED' | 'NO_SHOW' | 'WALK_OVER' | 'NO_PARTY' | string;

    participants: Participant[];

    [key: string]: any;
};